Create new meta table to link numer and geom timespans
We need a way to filter geometries based on the numerator timmespan.
Why? because it'd very useful to have a way to only show the geometries
that are going to have data for the selected numerator in the selected
timespan.

The current data only let us to filter the geometries by the geometry
timespan becuase the numerator timespan is aggregated.

This new table let us know for which timespans for a numerator and a
geometry have data.

For more information about this change, check this issue
https://github.com/CartoDB/bigmetadata/issues/178