Recursive generation of RSTs: supports 3+ levels
This changes the data structure for holding tags to generate
RST data to a tree with the different paths to reach a column.

Code is simplified by making it recursive, and solves the issue
with 3+ levels that the previous code had (where it rendered the
second level column in depths over 3).