Fix casing constraint existence check
On table creation, unless double quotes are used, table names are
lowercased. Because of that, if you use a table name with upper case,
even in the dictionary it's recorded as lowercased. So, unless you
explicitly downcase the table name for this check, you get a
nonexistence result. Instead of forcing downcasing, which might fail if
double quotes were used on creation, ilike makes this check looser.